improve controllers and services

+++ backend/src/services/WriteFarmService.ts
@@ -1,12 +1,19 @@
 import { Farm } from "~/entities/Farm";
 import IWriteFarmRepository from "~/contracts/IWriteFarmRepository";
 import IWriteFarmService from "~/contracts/IWriteFarmService";
+import IReadFarmRepository from "~/contracts/IReadFarmRepository";
+import createError from "http-errors";
 
 export class WriteFarmService implements IWriteFarmService {
   private writeFarmRepository: IWriteFarmRepository;
+  private readFarmRepository: IReadFarmRepository;
 
-  constructor(writeFarmRepository: IWriteFarmRepository) {
+  constructor(
+    writeFarmRepository: IWriteFarmRepository,
+    readFarmRepository: IReadFarmRepository
+  ) {
     this.writeFarmRepository = writeFarmRepository;
+    this.readFarmRepository = readFarmRepository;
   }
 
   async createFarm(farm: Farm): Promise<Farm> {
@@ -14,10 +21,20 @@ export class WriteFarmService implements IWriteFarmService {
   }
 
   async updateFarm(farm: Farm, id: number): Promise<Farm> {
+    const findFarm = await this.readFarmRepository.findOneById(id);
+    if (!findFarm) {
+      throw new createError.NotFound("Farm not found");
+    }
+
     return await this.writeFarmRepository.save(farm, id);
   }
 
   async deleteFarm(id: number): Promise<boolean> {
+    const findFarm = await this.readFarmRepository.findOneById(id);
+    if (!findFarm) {
+      throw new createError.NotFound("Farm not found");
+    }
+
     return await this.writeFarmRepository.delete(id);
   }
 }
